Transform도 Component로 취급하도록 되돌림(불편함)

+++ Project/Client/guiTexture.cpp
@@ -35,8 +35,8 @@ namespace gui
 		//ImGui::InputText("##TexPath"
 		//	, (char*)path.data(), path.size(), ImGuiInputTextFlags_ReadOnly);
 
-		int width = targetTex->GetWidth();
-		int height = targetTex->GetHeight();
+		int width = (int)targetTex->GetWidth();
+		int height = (int)targetTex->GetHeight();
 
 
 		char wBuffer[256] = "";
